requirements.txt
# 基础GUI依赖
PyQt5>=5.15.0
PyQt5-sip>=12.8.0

# Git相关依赖
GitPython>=3.1.0
gitdb>=4.0.5

# 网络请求
requests>=2.25.1
urllib3>=1.26.5

# 日期时间处理
python-dateutil>=2.8.2

# 文件处理
pathlib>=1.0.1

# 进度条
tqdm>=4.65.0

# 代码分析
ast>=0.0.2
astroid>=2.8.0

# 可选：代码高亮
Pygments>=2.10.0

# 系统相关
psutil>=5.8.0

# 编码处理
chardet>=4.0.0

# 可选：并发处理
concurrent-futures>=3.0.5

# 开发工具（可选）
pylint>=2.11.1
black>=21.9b0 